Refactor(src): move RAG-related functions to database module
